fastapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic[email]==2.7.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic
psycopg2-binary
